pyproject.toml
[tool.poetry]
name = "ai-mentor"
version = "0.1.0"
description = "Адаптивная образовательная платформа для школьников"
authors = ["AI Mentor Team"]
readme = "README.md"
packages = [{include = "app", from = "backend"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
gunicorn = "^23.0.0"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.0"
email-validator = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.3.0"
python-multipart = "^0.0.12"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
pgvector = "^0.3.5"
langchain = "^0.3.7"
langchain-openai = "^0.2.8"
openai = "^1.54.4"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"
google-auth = "^2.36.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
black = "^24.10.0"
ruff = "^0.7.4"
mypy = "^1.13.0"

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "-v --cov=backend/app --cov-report=html --cov-report=term"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
